pyproject.toml
[project]
name = "repl-nix-bi-an-shou-yi-tong-ji-gong-ju"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "python-binance>=1.0.22",
    "pytz>=2024.2",
    "streamlit>=1.39.0",
]
